pyproject.toml
[project]
name = "ai-minds"
version = "1.0.0"
description = "AI MINDS - Multimodal personal knowledge assistant"
requires-python = ">=3.9"
dependencies = [
    # Note: sqlite3 is built-in, no need to install
    # Browser extraction dependencies
    "browser-history>=0.3.2",
    # Clipboard watcher dependencies
    "pyperclip>=1.8.2",
    "Pillow>=10.0.0",
    "pywin32>=306 ; sys_platform == 'win32'",
    # Calendar watcher dependencies
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["Data Layer"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
